services/ai-processing/requirements.txt
fastapi==0.115.4
uvicorn==0.34.0
google-generativeai==0.8.3
xgboost==2.1.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
pydantic==2.9.2
python-dotenv==1.0.1
pydantic-settings==2.5.2

